generator/requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tensorboard>=2.14.0

# Shared moonboard utilities
-e ../moonboard_core

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

